Make sensor examples more versatile
