+++ circuitpython/sensors/intro/level3/sensor_internal.py
@@ -1,0 +1,57 @@
+"""
+Now, we will enhance the previous example and visualize the measured data on the display.
+
+It requires the adafruit_scd4x, adafruit_display_text and adafruit_imageload libraries placed in /lib directory.
+You can find the libraries in the CircuitPython library bundle (https://circuitpython.org/libraries).
+"""
+import microcontroller
+# Measurement class to store the sensor data in a structured way
+class Measurement:
+    def __init__(self, name, tag, type, value, unit):
+        # name: human-readable name of the measured value - e.g. Temperature
+        self.name = name
+        # tag: identifier of the measured value, usefull for data processing - e.g. temp1
+        self.tag = tag
+        # type: type of the measured value, useful for visualization - e.g. temperature
+        self.type = type
+        # value: the measured value itself - e.g. 25.6
+        self.value = value
+        # unit: unit of the measured value - e.g. °C
+        self.unit = unit
+
+    # basic string representation of the Measurement object
+    def __str__(self):
+        return "%s: %.1f %s" % (self.name, self.value, self.unit)
+
+# Sensor class to handle the sensor inicialization and data collection
+class Sensor:
+    def __init__(self):
+        # store the measured values in a list of Measurement objects
+        self.measurements = []
+        self.readings = {
+           "temp": {"name": "Temperature", "type": "temperature", "unit": "°C"},
+
+        }
+        
+
+    # measure the sensor data and return a list of Measurement objects
+    def measure(self):
+        self.measurements = []
+        
+        tag = "temp"
+        measurement = Measurement(
+            self.readings[tag]["name"], tag, self.readings[tag]["type"], microcontroller.cpu.temperature, self.readings[tag]["unit"])
+        self.measurements.append(measurement)
+
+        return self.measurements
+    # return the list of Measurement objects
+    def get_measurements(self):
+        return self.measurements
+    
+    # return only one Measurement object with the specified tag
+    def get_measurement(self, tag):
+        for measurement in self.measurements:
+            if measurement.tag == tag:
+                return measurement
+        return None
+
